合并 main，准备 Engram SSD CUDA graph 实验。

+++ infx/results/agentic/backends/base.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 from collections.abc import Iterable
 from typing import Any
 
-
 SERVER_CACHE_FLAT_FIELDS = {
     "server_gpu_cache_hit_rate": None,
     "server_cpu_cache_hit_rate": None,
@@ -41,8 +40,8 @@ class ServerMetricsBackend:
 
     def gpu_kv_capacity_tokens(
         self,
-        metrics: dict[str, dict[str, Any]],
-        server_logs: Iterable[str | None],
+        metrics: dict[str, dict[str, Any]],  # noqa: ARG002
+        server_logs: Iterable[str | None],  # noqa: ARG002
     ) -> int | None:
         return None
 
@@ -111,7 +110,7 @@ def apply_profile_totals(flat: dict[str, Any], records: list[dict[str, Any]]) ->
 def counter_int(value: float | None) -> int | None:
     if value is None:
         return None
-    return int(round(value))
+    return round(value)
 
 
 def _record_token_sum(records: list[dict[str, Any]], metric_name: str) -> int | None:
